Add custom resize handler to IconBox class
A custom resize handler has been added to the IconBox class. This handler adjusts the Top position and Height of iconListView whenever the form is resized. Specifically, it sets the Top position of iconListView to HEADER_HEIGHT and calculates the available height for iconListView by subtracting HEADER_HEIGHT from the form's client height.